Osa 4 t4.18-4.21 valmis

+++ controllers/blogs.js
@@ -1,5 +1,8 @@
+const jwt = require('jsonwebtoken')
 const router = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
 
 router.get('/', async (req, res) => {
   const blogs = await Blog.find({})
@@ -17,24 +20,47 @@ router.get('/:id', async (req, res) => {
     } else {
       res.status(404).end()
     }
-  } catch(exception) {
+  } catch(error) {
     // console.log(exception)
     res.status(400).send({ error: 'malformatted id' })
   }
 })
 
 router.post('/', async (req, res) => {
-  const newBlogData = req.body
-  if(!newBlogData.title || !newBlogData.url) {
-    return res.status(400).json({ error: 'content missing' })
-  }
+  try {
+    const token = req.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
 
-  if(!newBlogData.likes) {
-    newBlogData.likes = 0
+    if(!token || !decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const newBlogData = req.body
+    if(!newBlogData.title || !newBlogData.url) {
+      return res.status(400).json({ error: 'content missing' })
+    }
+
+    if(!newBlogData.likes) {
+      newBlogData.likes = 0
+    }
+    newBlogData.user = decodedToken.id
+    const blog = new Blog(newBlogData)
+    const result = await blog.save()
+
+    const user = await User.findById(decodedToken.id)
+    user.blogs = user.blogs.concat(result._id)
+    await user.save()
+
+    res.status(201).json(Blog.format(result))
+
+  } catch(error) {
+    if(error.name === 'JsonWebTokenError') {
+      res.status(401).json({ error: error.message })
+    } else {
+      console.log(error)
+      res.status(500).json({ error: 'something went wrong...' })
+    }
   }
-  const blog = new Blog(newBlogData)
-  const result = await blog.save()
-  res.status(201).json(Blog.format(result))
 })
 
 router.put('/:id', async (req, res) => {
@@ -54,9 +80,21 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
   try {
+    const token = req.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if(!token || !decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(req.params.id)
+    if(blog.user.toString() !== decodedToken.id) {
+      return res.status(401).json({ error: 'not allowed to remove blog' })
+    }
     await Blog.findByIdAndRemove(req.params.id)
     res.status(204).end()
   } catch(error) {
+    console.error(error)
     res.status(400).send({ error: 'malformatted id' })
   }
 })
